added style too display allergies and also added buttons that link back to display medications and display medical probs also deleted a random php file that was inside the APP folder
